+++ src/Guzzle/Http/Message/Request.php
@@ -11,12 +11,12 @@ use Guzzle\Http\Exception\RequestException;
 use Guzzle\Http\Exception\BadResponseException;
 use Guzzle\Http\ClientInterface;
 use Guzzle\Http\QueryString;
-use Guzzle\Http\Cookie;
 use Guzzle\Http\EntityBody;
 use Guzzle\Http\Url;
 
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
  * HTTP request class to send requests
@@ -57,11 +57,6 @@ class Request extends AbstractMessage implements RequestInterface
      * @var string State of the request object
      */
     protected $state;
-
-    /**
-     * @var Cookie Cookies to send with the request
-     */
-    protected $cookie;
 
     /**
      * @param string Auth username
@@ -148,7 +143,6 @@ class Request extends AbstractMessage implements RequestInterface
             $this->setHeader('User-Agent', Utils::getDefaultUserAgent());
         }
 
-        $this->cookie = Cookie::factory($this->getHeader('Cookie'));
         $this->setState(self::STATE_NEW);
     }
 
@@ -238,7 +232,7 @@ class Request extends AbstractMessage implements RequestInterface
      */
     public function getRawHeaders()
     {
-        $raw = $this->method . ' ' . $this->getResourceUri();
+        $raw = $this->method . ' ' . $this->getResource();
         $protocolVersion = $this->protocolVersion ?: '1.1';
         $raw = trim($raw) . ' ' . strtoupper(str_replace('https', 'http', $this->url->getScheme())) . '/' . $protocolVersion . "\r\n";
         $raw .= $this->getHeaderString();
@@ -281,7 +275,7 @@ class Request extends AbstractMessage implements RequestInterface
      * Send the request
      *
      * @return Response
-     * @throws RequestException on a request error
+     * @throws RuntimeException if a client is not associated with the request
      */
     public function send()
     {
@@ -488,7 +482,6 @@ class Request extends AbstractMessage implements RequestInterface
      * @return Request
      *
      * @see http://www.ietf.org/rfc/rfc2617.txt
-     * @throws RequestException
      */
     public function setAuth($user, $password = '', $scheme = CURLAUTH_BASIC)
     {
@@ -514,12 +507,12 @@ class Request extends AbstractMessage implements RequestInterface
     }
 
     /**
-     * Get the URI of the request (e.g. '/', '/index.html', '/index.html?q=1)
-     * This is the path plus the query string, fragment
+     * Get the resource part of the the request, including the path, query
+     * string, and fragment
      *
      * @return string
      */
-    public function getResourceUri()
+    public function getResource()
     {
         return $this->url->getPath() . (string) $this->url->getQuery();
     }
@@ -586,16 +579,12 @@ class Request extends AbstractMessage implements RequestInterface
      */
     public function receiveResponseHeader($data)
     {
+        static $normalize = array("\r", "\n");
         $this->state = self::STATE_TRANSFER;
         $length = strlen($data);
-        $data = str_replace(array("\r", "\n"), '', $data);
+        $data = str_replace($normalize, '', $data);
 
-        if (strpos($data, ':') !== false) {
-
-            list($header, $value) = explode(':', $data, 2);
-            $this->response->addHeader(trim($header), trim($value));
-
-        } elseif (strlen($data) > 6) {
+        if (strpos($data, 'HTTP/') === 0) {
 
             list($dummy, $code, $status) = explode(' ', $data, 3);
 
@@ -616,6 +605,11 @@ class Request extends AbstractMessage implements RequestInterface
                 'reason_phrase'     => $status,
                 'previous_response' => $previousResponse
             ));
+
+        } elseif (strpos($data, ':') !== false) {
+
+            list($header, $value) = explode(':', $data, 2);
+            $this->response->addHeader(trim($header), trim($value));
         }
 
         return $length;
@@ -682,41 +676,35 @@ class Request extends AbstractMessage implements RequestInterface
     }
 
     /**
-     * Get an array of Cookies or a specific cookie from the request
+     * Get an array of cookies
      *
-     * @param string $name Cookie to retrieve
-     *
-     * @return null|string|Cookie Returns null if not found by name, a Cookie
-     *      object if no $name is supplied, or the cookie value by name if found
-     *      If a Cookie object is returned, changes to the cookie object does
-     *      not modify the request's cookies.  You will need to set the cookie
-     *      back on the request after modifying the object.
+     * @return array
      */
-    public function getCookie($name = null)
+    public function getCookies()
     {
-        return !$name ? clone $this->cookie : $this->cookie->get($name);
+        $cookieData = new Collection();
+        if ($cookies = $this->getHeader('Cookie')) {
+            foreach ($cookies as $cookie) {
+                $parts = explode('=', $cookie, 2);
+                $cookieData->add($parts[0], isset($parts[1]) ? $parts[1] : '');
+            }
+        }
+
+        return $cookieData->getAll();
     }
 
     /**
-     * Set the Cookie header using an array or Cookie object
+     * Get a cookie value by name
      *
-     * @param array|Cookie $cookies Cookie data to set on the request
+     * @param string $name Cookie to retrieve
      *
-     * @return Request
+     * @return null|string|array
      */
-    public function setCookie($cookies)
+    public function getCookie($name)
     {
-        if ($cookies instanceof Cookie) {
-            $this->cookie = $cookies;
-        } elseif (is_array($cookies)) {
-            $this->cookie->replace($cookies);
-        } else {
-            throw new InvalidArgumentException('Invalid cookie data');
-        }
+        $cookies = $this->getCookies();
 
-        $this->headers['cookie'] = new Header('Cookie', (string) $this->cookie);
-
-        return $this;
+        return isset($cookies[$name]) ? $cookies[$name] : null;
     }
 
     /**
@@ -729,27 +717,35 @@ class Request extends AbstractMessage implements RequestInterface
      */
     public function addCookie($name, $value)
     {
-        $this->cookie->add($name, $value);
-        $this->headers['cookie'] = new Header('Cookie', (string) $this->cookie);
+        if (!$this->hasHeader('Cookie')) {
+            $this->setHeader('Cookie', "{$name}={$value}");
+        } else {
+            $this->getHeader('Cookie')->add("{$name}={$value}");
+        }
 
         return $this;
     }
 
     /**
-     * Remove the cookie header or a specific cookie value by name
+     * Remove a specific cookie value by name
      *
-     * @param string $name Cookie to remove by name.  If no value is
-     *      provided, the entire Cookie header is removed from the request
+     * @param string $name Cookie to remove by name
      *
      * @return Request
      */
-    public function removeCookie($name = null)
+    public function removeCookie($name)
     {
-        $this->cookie->remove($name);
-        $this->headers['cookie'] = new Header('Cookie', (string) $this->cookie);
+        if ($cookie = $this->getHeader('Cookie')) {
+            foreach ($cookie as $cookieValue) {
+                if (strpos($cookieValue, $name . '=') === 0) {
+                    $cookie->removeValue($cookieValue);
+                }
+            }
+        }
 
         return $this;
     }
+
 
     /**
      * Returns whether or not the response served to the request can be cached
@@ -805,6 +801,17 @@ class Request extends AbstractMessage implements RequestInterface
 
     /**
      * {@inheritdoc}
+     * @codeCoverageIgnore
+     */
+    public function addSubscriber(EventSubscriberInterface $subscriber)
+    {
+        $this->getEventDispatcher()->addSubscriber($subscriber);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
      */
     protected function changedHeader($action, $header)
     {
@@ -813,13 +820,6 @@ class Request extends AbstractMessage implements RequestInterface
         if ($header === 'host') {
             // If the Host header was changed, be sure to update the internal URL
             $this->setHost((string) $this->getHeader('Host'));
-        } elseif ($header === 'cookie') {
-            // Be sure to get an cookie updates and update the internal Cookie
-            if ($action === 'set') {
-                $this->cookie = Cookie::factory($this->getHeader('Cookie'));
-            } elseif ($this->cookie) {
-                $this->cookie->clear();
-            }
         }
     }