+++ tests/Guzzle/Tests/Http/EntityBodyTest.php
@@ -3,6 +3,7 @@
 namespace Guzzle\Tests\Http;
 
 use Guzzle\Http\EntityBody;
+use Guzzle\Http\QueryString;
 
 /**
  * @group server
@@ -20,6 +21,7 @@ class EntityBodyTest extends \Guzzle\Tests\GuzzleTestCase
 
     /**
      * @covers Guzzle\Http\EntityBody::factory
+     * @covers Guzzle\Http\EntityBody::fromString
      */
     public function testFactory()
     {
@@ -29,15 +31,15 @@ class EntityBodyTest extends \Guzzle\Tests\GuzzleTestCase
         $this->assertEquals('php', $body->getWrapper());
         $this->assertEquals('temp', $body->getStreamType());
 
-        $handle = fopen(__DIR__ . '/../../../../phpunit.xml', 'r');
+        $handle = fopen(__DIR__ . '/../../../../phpunit.xml.dist', 'r');
         if (!$handle) {
             $this->fail('Could not open test file');
         }
         $body = EntityBody::factory($handle);
-        $this->assertEquals(__DIR__ . '/../../../../phpunit.xml', $body->getUri());
+        $this->assertEquals(__DIR__ . '/../../../../phpunit.xml.dist', $body->getUri());
         $this->assertTrue($body->isLocal());
-        $this->assertEquals(__DIR__ . '/../../../../phpunit.xml', $body->getUri());
-        $this->assertEquals(filesize(__DIR__ . '/../../../../phpunit.xml'), $body->getContentLength());
+        $this->assertEquals(__DIR__ . '/../../../../phpunit.xml.dist', $body->getUri());
+        $this->assertEquals(filesize(__DIR__ . '/../../../../phpunit.xml.dist'), $body->getContentLength());
 
         // make sure that a body will return as the same object
         $this->assertTrue($body === EntityBody::factory($body));
@@ -54,6 +56,24 @@ class EntityBodyTest extends \Guzzle\Tests\GuzzleTestCase
         $body = EntityBody::factory();
         $this->assertEquals('php', $body->getWrapper());
         $this->assertEquals('temp', $body->getStreamType());
+    }
+
+    /**
+     * @covers Guzzle\Http\EntityBody::factory
+     */
+    public function testFactoryCanCreateFromObject()
+    {
+        $body = EntityBody::factory(new QueryString(array('foo' => 'bar')));
+        $this->assertEquals('?foo=bar', (string) $body);
+    }
+
+    /**
+     * @covers Guzzle\Http\EntityBody::factory
+     * @expectedException Guzzle\Common\Exception\InvalidArgumentException
+     */
+    public function testFactoryEnsuresObjectsHaveToStringMethod()
+    {
+        EntityBody::factory(new \stdClass('a'));
     }
 
     /**