+++ tests/Guzzle/Tests/Service/ClientTest.php
@@ -2,16 +2,15 @@
 
 namespace Guzzle\Tests\Service;
 
-use Guzzle\Common\Guzzle;
 use Guzzle\Common\Collection;
 use Guzzle\Common\Log\ClosureLogAdapter;
+use Guzzle\Inflection\Inflector;
 use Guzzle\Http\Message\Response;
 use Guzzle\Http\Message\RequestFactory;
 use Guzzle\Http\Curl\CurlMulti;
 use Guzzle\Http\Plugin\MockPlugin;
 use Guzzle\Service\Description\ApiCommand;
 use Guzzle\Service\Client;
-use Guzzle\Service\Command\CommandSet;
 use Guzzle\Service\Command\CommandInterface;
 use Guzzle\Service\Description\XmlDescriptionBuilder;
 use Guzzle\Service\Description\ServiceDescription;
@@ -108,55 +107,12 @@ class ClientTest extends \Guzzle\Tests\GuzzleTestCase
 
     /**
      * @covers Guzzle\Service\Client::execute
-     * @expectedException Guzzle\Service\Exception\CommandSetException
-     */
-    public function testThrowsExceptionWhenExecutingMixedClientCommandSets()
-    {
-        $client = new Client('http://www.test.com/');
-        $otherClient = new Client('http://www.test-123.com/');
-
-        // Create a command set and a command
-        $set = new CommandSet();
-        $cmd = new MockCommand();
-        $set->addCommand($cmd);
-
-        // Associate the other client with the command
-        $cmd->setClient($otherClient);
-
-        // Send the set with the wrong client, causing an exception
-        $client->execute($set);
-    }
-
-    /**
-     * @covers Guzzle\Service\Client::execute
      * @expectedException Guzzle\Common\Exception\InvalidArgumentException
      */
-    public function testThrowsExceptionWhenExecutingInvalidCommandSets()
+    public function testThrowsExceptionWhenInvalidCommandIsExecuted()
     {
-        $client = new Client('http://www.test.com/');
+        $client = new Client();
         $client->execute(new \stdClass());
-    }
-
-    /**
-     * @covers Guzzle\Service\Client::execute
-     */
-    public function testExecutesCommandSets()
-    {
-        $client = new Client('http://www.test.com/');
-        $client->getEventDispatcher()->addSubscriber(new MockPlugin(array(
-            new Response(200)
-        )));
-
-        // Create a command set and a command
-        $set = new CommandSet();
-        $cmd = new MockCommand();
-        $set->addCommand($cmd);
-        $this->assertSame($set, $client->execute($set));
-
-        // Make sure it sent
-        $this->assertTrue($cmd->isExecuted());
-        $this->assertTrue($cmd->isPrepared());
-        $this->assertEquals(200, $cmd->getResponse()->getStatusCode());
     }
 
     /**
@@ -337,21 +293,6 @@ class ClientTest extends \Guzzle\Tests\GuzzleTestCase
     }
 
     /**
-     * @covers Guzzle\Service\Client::getCommand
-     * @depends testMagicCallBehaviorExecuteExecutesCommands
-     */
-    public function testEnablesMagicMethodCallsOnCommandsIfEnabledOnClient()
-    {
-        $client = new Mock\MockClient();
-        $command = $client->getCommand('other_command');
-        $this->assertNull($command->get('command.magic_method_call'));
-
-        $client->setMagicCallBehavior(Client::MAGIC_CALL_EXECUTE);
-        $command = $client->getCommand('other_command');
-        $this->assertTrue($command->get('command.magic_method_call'));
-    }
-
-    /**
      * @covers Guzzle\Service\Client::execute
      */
     public function testClientResetsRequestsBeforeExecutingCommands()
@@ -405,11 +346,25 @@ class ClientTest extends \Guzzle\Tests\GuzzleTestCase
      */
     public function testClientCreatesIteratorsWithCommands()
     {
-       $client = new Mock\MockClient();
-       $command = new MockCommand();
-       $iterator = $client->getIterator($command);
-       $this->assertInstanceOf('Guzzle\Tests\Service\Mock\Model\MockCommandIterator', $iterator);
-       $iteratorCommand = $this->readAttribute($iterator, 'originalCommand');
-       $this->assertSame($command, $iteratorCommand);
-   }
+        $client = new Mock\MockClient();
+        $command = new MockCommand();
+        $iterator = $client->getIterator($command);
+        $this->assertInstanceOf('Guzzle\Tests\Service\Mock\Model\MockCommandIterator', $iterator);
+        $iteratorCommand = $this->readAttribute($iterator, 'originalCommand');
+        $this->assertSame($command, $iteratorCommand);
+    }
+
+    /**
+     * @covers Guzzle\Service\Client::getInflector
+     * @covers Guzzle\Service\Client::setInflector
+     */
+    public function testClientHoldsInflector()
+    {
+        $client = new Mock\MockClient();
+        $this->assertInstanceOf('Guzzle\Inflection\MemoizingInflector', $client->getInflector());
+
+        $inflector = new Inflector();
+        $client->setInflector($inflector);
+        $this->assertSame($inflector, $client->getInflector());
+    }
 }