+++ tests/Guzzle/Tests/Service/Description/ApiCommandTest.php
@@ -35,36 +35,46 @@ class ApiCommandTest extends \Guzzle\Tests\GuzzleTestCase
     public function testApiCommandIsDataObject()
     {
         $c = new ApiCommand(array(
-            'name' => 'test',
-            'doc' => 'doc',
-            'method' => 'POST',
-            'uri' => '/api/v1',
+            'name'    => 'test',
+            'doc'     => 'doc',
+            'doc_url' => 'http://www.example.com',
+            'method'  => 'POST',
+            'uri'     => '/api/v1',
+            'result_type' => 'array',
+            'result_doc'  => 'returns the json_decoded response',
+            'deprecated'  => true,
             'params' => array(
                 'key' => array(
-                    'required' => 'true',
-                    'type' => 'string',
+                    'required'   => 'true',
+                    'type'       => 'string',
                     'max_length' => 10
                 ),
                 'key_2' => array(
                     'required' => 'true',
-                    'type' => 'integer',
-                    'default' => 10
+                    'type'     => 'integer',
+                    'default'  => 10
                 )
            )
         ));
 
         $this->assertEquals('test', $c->getName());
         $this->assertEquals('doc', $c->getDoc());
+        $this->assertEquals('http://www.example.com', $c->getDocUrl());
         $this->assertEquals('POST', $c->getMethod());
         $this->assertEquals('/api/v1', $c->getUri());
+        $this->assertEquals('array', $c->getResultType());
+        $this->assertEquals('returns the json_decoded response', $c->getResultDoc());
+        $this->assertTrue($c->isDeprecated());
         $this->assertEquals('Guzzle\\Service\\Command\\DynamicCommand', $c->getConcreteClass());
         $this->assertEquals(array(
             'key' => new ApiParam(array(
+                'name' => 'key',
                 'required' => 'true',
                 'type' => 'string',
                 'max_length' => 10
             )),
             'key_2' => new ApiParam(array(
+                'name' => 'key_2',
                 'required' => 'true',
                 'type' => 'integer',
                 'default' => 10
@@ -72,6 +82,7 @@ class ApiCommandTest extends \Guzzle\Tests\GuzzleTestCase
         ), $c->getParams());
 
         $this->assertEquals(new ApiParam(array(
+            'name' => 'key_2',
             'required' => 'true',
             'type' => 'integer',
             'default' => 10
@@ -103,16 +114,20 @@ class ApiCommandTest extends \Guzzle\Tests\GuzzleTestCase
     public function testConvertsToArray()
     {
         $data = array(
-            'name' => 'test',
+            'name'      => 'test',
             'class'     => 'Guzzle\\Service\\Command\ClosureCommand',
             'doc'       => 'test',
+            'doc_url'   => 'http://www.example.com',
             'method'    => 'PUT',
             'uri'       => '/',
             'params'    => array(
                 'p' => new ApiParam(array(
                     'name' => 'foo'
                 ))
-            )
+            ),
+            'result_type' => null,
+            'result_doc'  => null,
+            'deprecated'  => false
         );
         $c = new ApiCommand($data);
         $this->assertEquals($data, $c->toArray());